fix: exporting games for user by updating structs
Merge request #41 changed GameJson - now causing exports to fail.
Originally ChallengeAI used GameJson and it was modified accordingly.
This was incorrect however, and I changed ChallengeAI to use Move.
However I failed to revert the changes to GameJson.
This wasn't seen because GameJson isn't used in many places.

In addition, I've remove any null title from the test files.
I'm assuming this was a bug in an older version of lila.
The newer game_json example omits title if null, as it's supposed to.
Therefore I'm assuming this is the expected behaviour.